requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
arxiv==2.2.0
asyncio-mqtt==0.16.2
attrs==25.4.0
backoff==2.2.1
beautifulsoup4==4.14.2
black==25.9.0
Brotli==1.1.0
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
contourpy==1.3.3
cryptography==46.0.2
cycler==0.12.1
dataclasses-json==0.6.7
deprecation==2.1.0
distro==1.9.0
et_xmlfile==2.0.0
eval_type_backport==0.2.2
fastapi==0.119.0
feedparser==6.0.12
ffmpy==0.6.3
filelock==3.20.0
filetype==1.2.0
flake8==7.3.0
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2025.9.0
gigachat==0.2.1
google-ai-generativelanguage==0.6.15
google-api-core==2.26.0
google-api-python-client==2.184.0
google-auth==2.50.0
google-auth-httplib2==0.2.0
google-genai==1.74.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
gradio>=6.4.0
gradio_client>=2.0.3
greenlet==3.2.4
groovy==0.1.2
groq==0.32.0
grpcio==1.75.1
grpcio-status>=1.49.1,<2.0.0
h11==0.16.0
h2==4.3.0
hf-xet==1.1.10
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.35.3
hyperframe==6.1.0
idna==3.11
importlib_metadata==8.7.0
iniconfig==2.1.0
invoke==2.2.1
Jinja2==3.1.6
jiter==0.11.0
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.9
langchain==1.2.17
langchain-classic==1.0.4
langchain-community==0.4.1
langchain-core==1.3.2
langchain-gigachat==0.5.0
langchain-google-genai==4.2.2
langchain-groq==1.1.2
langchain-huggingface==1.2.2
langchain-mistralai==1.1.3
langchain-openai==1.2.1
langchain-protocol==0.0.15
langchain-tavily==0.2.18
langchain-text-splitters==1.1.2
langfuse==3.6.2
langgraph==1.1.10
langgraph-checkpoint==4.0.3
langgraph-prebuilt==1.0.13
langgraph-sdk==0.3.13
langsmith==0.4.34
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.1
matplotlib==3.10.7
mccabe==0.7.0
mdurl==0.1.2
mistralai==1.9.11
multidict==6.7.0
mypy_extensions==1.1.0
numpy==2.3.3
openai==2.33.0
openpyxl==3.1.5
opentelemetry-api==1.37.0
opentelemetry-exporter-otlp-proto-common==1.37.0
opentelemetry-exporter-otlp-proto-http==1.37.0
opentelemetry-proto==1.37.0
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0
orjson==3.11.8
ormsgpack==1.12.2
packaging==25.0
paho-mqtt==2.1.0
pandas==2.3.3
pathspec==0.12.1
pgvector==0.4.1
pillow==11.3.0
platformdirs==4.5.0
pluggy==1.6.0
# Optional Supabase client stack — not used by agent_ng/ or tools/ (legacy .legacy/agent_old only).
# Uncomment if you need SUPABASE_URL workflows; then: pip install -r requirements.txt
# postgrest==2.29.0
# realtime==2.29.0
# storage3==2.29.0
# supabase==2.29.0
# supabase-auth==2.29.0
# supabase-functions==2.29.0
propcache==0.4.1
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.14.0
pycparser==2.23
pydantic==2.11.10
pydantic-settings==2.11.0
pydantic_core==2.33.2
pydub==0.25.1
pyflakes==3.4.0
Pygments==2.19.2
PyJWT==2.10.1
PyMuPDF==1.26.5
pymupdf4llm==0.0.27
pyparsing==3.2.5
pytest==8.4.2
pytest-asyncio==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytokens==0.1.10
pytz==2025.2
PyYAML==6.0.3
regex==2025.9.18
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.2.0
rsa==4.9.1
ruff==0.14.0
safehttpx==0.1.6
scikit-learn==1.7.2
scipy==1.16.2
semantic-version==2.10.0
setuptools==80.9.0
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.44
starlette==0.48.0
StrEnum==0.4.15
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.1
tomlkit==0.13.3
tqdm==4.67.1
typer==0.19.2
types-requests==2.32.4.20250913
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uuid_utils==0.14.1
uvicorn==0.37.0
websockets==15.0.1
markitdown[docx,xlsx,pptx]
wheel==0.45.1
wikipedia==1.4.0
wrapt==1.17.3
xxhash==3.7.0
yarl==1.22.0
zipp==3.23.0
zstandard==0.25.0

